Move shared aux functions for download to separate submodule
